refactor: move track to its own component
